pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Discord bot backed by Grok"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "python-dotenv>=1.2.1",
    "xai-sdk>=1.6.0",
]

[project.scripts]
bot = "bot.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
python_version = "3.13"

[[tool.mypy.overrides]]
module = [
    "xai",
    "xai.*",
    "xai.api.v1",
    "xai.api.v1.*",
    "xai_sdk",
    "xai_sdk.*",
]
ignore_missing_imports = true
